Implement snapshots

+++ meta/store_test.go
@@ -608,6 +608,54 @@ func TestStore_UserCount(t *testing.T) {
 	}
 }
 
+// Ensure the store can take a snapshot.
+func TestStore_Snapshot_And_Restore(t *testing.T) {
+	t.Parallel()
+
+	s := MustOpenStore()
+	s.LeaveFiles = true
+
+	// Create a bunch of databases in the Store
+	nDatabases := 5
+	for n := 0; n < nDatabases; n++ {
+		s.CreateDatabase(fmt.Sprintf("db%d", n))
+	}
+
+	// Test taking a snapshot.
+	if err := s.Store.Snapshot(); err != nil {
+		t.Fatal(err)
+	}
+
+	s.Close()
+
+	// Test restoring the snapshot taken above.
+	existingDataPath := s.Path()
+	s = NewStore(NewConfig(existingDataPath))
+	if err := s.Open(); err != nil {
+		panic(err)
+	}
+	defer s.Close()
+
+	// Wait until the server is ready.
+	select {
+	case err := <-s.Err():
+		panic(err)
+	case <-s.Ready():
+	}
+
+	// Make sure all the data we added to the Store is still there.
+	for n := 0; n < nDatabases; n++ {
+		name := fmt.Sprintf("db%d", n)
+		if dbi, err := s.Database(name); err != nil {
+			t.Fatal(err)
+		} else if dbi == nil {
+			t.Fatalf("database not found: %s", name)
+		} else if dbi.Name != name {
+			t.Fatal(name)
+		}
+	}
+}
+
 // Ensure a multi-node cluster can start, join the cluster, and replicate commands.
 func TestCluster_Open(t *testing.T) {
 	c := MustOpenCluster(3)
@@ -642,8 +690,9 @@ func TestCluster_Open(t *testing.T) {
 // Store is a test wrapper for meta.Store.
 type Store struct {
 	*meta.Store
-	Listener net.Listener
-	Stderr   bytes.Buffer
+	Listener   net.Listener
+	Stderr     bytes.Buffer
+	LeaveFiles bool // set to true to leave temporary files on close
 }
 
 // NewStore returns a new test wrapper for Store.
@@ -701,7 +750,9 @@ func (s *Store) Close() error {
 	if s.Listener != nil {
 		s.Listener.Close()
 	}
-	defer os.RemoveAll(s.Path())
+	if !s.LeaveFiles {
+		defer os.RemoveAll(s.Path())
+	}
 	return s.Store.Close()
 }
